refactor(e2e): unify setup and setupFromFresh into single function
Consolidate the two e2e test setup functions into a unified setup function
in setup.ts. Both utils.ts and snapshot_manager.ts now re-export from setup.ts
with backward-compatible wrappers.

Key changes:
- Create setup.ts with unified EndToEndContext type and setup function
- Add aztecNodeConfig alias for backward compatibility with SubsystemsContext
- snapshot_manager.ts provides SubsystemsContext type (with non-optional fields)
  and setupFromFresh wrapper function
- utils.ts re-exports from setup.ts for full backward compatibility
- Both fundSponsoredFPC and skipAccountDeployment flags added to SetupOptions
