+++ yarn-project/end-to-end/src/e2e_epochs/epochs_test.ts
@@ -211,7 +211,7 @@ export class EpochsTestContext {
           ...opts,
         },
         this.context.aztecNode,
-        undefined,
+        this.context.prefilledPublicData ?? [],
         { dateProvider: this.context.dateProvider },
       ),
     );
